move promtVariable(...) to a different file
